Create chungus package
