+++ src/main/java/chungus/Chungus.java
@@ -1,3 +1,5 @@
+package chungus;
+
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
