+++ src/main/java/chungus/Storage.java
@@ -1,3 +1,5 @@
+package chungus;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
