+++ src/main/java/chungus/Ui.java
@@ -1,3 +1,5 @@
+package chungus;
+
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
